+++ src/main/java/me/gfumagali/btgpedidos/repository/OrderTotalValueRepository.java
@@ -3,5 +3,5 @@ package me.gfumagali.btgpedidos.repository;
 import me.gfumagali.btgpedidos.repository.model.OrderTotalValue;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface OrderTotalValueRepository extends MongoRepository<OrderTotalValue, Integer> {
+public interface OrderTotalValueRepository extends MongoRepository<OrderTotalValue, Long> {
 }
